Move module to subdirectories and adjust use
