+++ crates/core/tedge_mapper/src/c8y/mod.rs
@@ -1,7 +1,10 @@
+pub mod converter;
 pub mod error;
-mod http_proxy;
+mod fragments;
+pub mod http_proxy;
 mod json_c8y;
 pub mod mapper;
+mod operations;
 pub mod topic;
 
 #[cfg(test)]
