+++ crates/core/tedge_mapper/src/collectd/error.rs
@@ -6,7 +6,7 @@ pub enum DeviceMonitorError {
     FromMqttClient(#[from] mqtt_channel::MqttError),
 
     #[error(transparent)]
-    FromInvalidCollectdMeasurement(#[from] crate::collectd_mapper::collectd::CollectdError),
+    FromInvalidCollectdMeasurement(#[from] crate::collectd::collectd::CollectdError),
 
     #[error(transparent)]
     FromInvalidThinEdgeJson(#[from] thin_edge_json::group::MeasurementGrouperError),
