+++ crates/core/tedge_mapper/src/mapping/component.rs
