+++ crates/core/tedge_mapper/src/mapping/operations.rs
@@ -6,7 +6,7 @@ use std::{
 
 use serde::Deserialize;
 
-use crate::error::OperationsError;
+use crate::mapping::error::OperationsError;
 
 /// Operations are derived by reading files subdirectories per cloud /etc/tedge/operations directory
 /// Each operation is a file name in one of the subdirectories
