+++ crates/core/tedge_mapper/src/mapping/size_threshold.rs
@@ -1,3 +1,4 @@
+#[derive(Debug)]
 pub struct SizeThreshold(pub usize);
 
 impl SizeThreshold {
